third_party/blink/web_tests/css3/calc/calc-nesting.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../css-parser/resources/property-parsing-test.js"></script>

<style>
#parent { width: 200px; }
#div1 { width: calc(calc(50px)); }
#div2 { width: calc(calc(60%) - 20px); }
#div3 { width: calc(calc(3 * 25%)); }
#div4 {
    --width: calc(10% + 30px);
    width: calc(2 * var(--width));
}
</style>
<div id=parent>
  <div id=div1></div>
  <div id=div2></div>
  <div id=div3></div>
  <div id=div4></div>
</div>

<script>
// Tests using property-parsing-test.js
assert_valid_value("left", "calc(20px + calc(80px))", "calc(100px)");
assert_valid_value("left", "calc(calc(100px))", "calc(100px)");
assert_valid_value("left", "calc(calc(2) * calc(50px)", "calc(100px)");
assert_valid_value("left", "calc(calc(150px*2/3)", "calc(100px)");
assert_valid_value("left", "calc(calc(2 * calc(calc(3)) + 4) * 10px)", "calc(100px)");
assert_valid_value("left", "calc(50px + calc(40%))", "calc(40% + 50px)");
assert_valid_value("border", "calc(calc(10px)) solid pink", "calc(10px) solid pink");

// Tests that require layout
test(function(){
    assert_equals(getComputedStyle(div1).width, "50px");
    assert_equals(getComputedStyle(div2).width, "100px");
    assert_equals(getComputedStyle(div3).width, "150px");
    assert_equals(getComputedStyle(div4).width, "100px");
}, "Nested calcs should work with layout");
</script>
